+++ src/lib.rs
@@ -8,3 +8,5 @@ pub(crate) mod signable;
 pub mod signer;
 /// Signature verification for incoming requests
 pub mod verifier;
+
+mod pem_format;