+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.6.0
+- Normalize provided private key string to the correct PEM format.
+- Update thiserror to version 2.
+
+# 0.5.0
+- Convert from anyhow to thiserror.
+
 # 0.4.1
 - Add the Debug and Clone traits to `Signer` and `Verifier`.